+++ source/engine/4chan/thread/parseThreadResponse.js
@@ -3,11 +3,12 @@ import parseThread from './parseThread'
 /**
  * Parses "get thread comments" API response.
  * @param  {object} response — "get thread comments" API response.
- * @return {object} `{ thread, comments }`
+ * @return {object} `{ thread }`
  */
 export default function parseThreadResponse(response) {
+	const thread = parseThread(response.posts[0])
+	thread.comments = thread.comments.concat(response.posts.slice(1))
 	return {
-		thread: parseThread(response.posts[0]),
-		comments: response.posts
+		thread
 	}
 }