+++ source/engine/lynxchan/thread/parseThreadResponse.js
@@ -4,12 +4,11 @@ import getBoardInfo from '../board/getBoardInfo'
 /**
  * Parses "get thread comments" API response.
  * @param  {object} response — "get thread comments" API response.
- * @return {object} `{ thread, comments, board }`
+ * @return {object} `{ thread, board }`
  */
 export default function parseThreadResponse(response) {
 	return {
-		thread: parseThread(response),
-		comments: [response].concat(response.posts || []),
+		thread: parseThread(response, { mode: 'thread' }),
 		board: getBoardInfo(response)
 	}
 }