+++ source/engine/lynxchan/thread/parseThreadsResponse.js
@@ -3,11 +3,10 @@ import parseThread from './parseThread'
 /**
  * Parses "get threads list" API response.
  * @param  {object} response — "get threads list" API response.
- * @return {object} `{ threads, comments }`
+ * @return {object} `{ threads }`
  */
 export default function parseThreadsResponse(response) {
 	return {
-		threads: response.map(parseThread),
-		comments: response
+		threads: response.map(thread => parseThread(thread, { mode: 'catalog' }))
 	}
 }