+++ source/engine/makaba/index.js
@@ -2,10 +2,12 @@ import Engine from '../../Engine'
 
 import parseBoardsResponse from './board/parseBoardsResponse'
 import parseThreadsResponse from './thread/parseThreadsResponse'
+import parseThreadsPageResponse from './thread/parseThreadsPageResponse'
 import parseThreadResponse from './thread/parseThreadResponse'
 import parseComment from './comment/parseComment'
 import parseVoteResponse from './vote/parseVoteResponse'
 import parsePostResponse from './post/parsePostResponse'
+import parseReportResponse from './report/parseReportResponse'
 
 import Board from '../../Board'
 import Thread from '../../Thread'
@@ -38,17 +40,19 @@ export default class Makaba extends Engine {
 	 * @return {Thread[]}
 	 */
 	parseThreads(response, options) {
-		const {
-			threads,
-			comments,
-			board
-		} = parseThreadsResponse(response)
-		return threads.map((thread, i) => Thread(
-			thread,
-			[this.parseComment(comments[i], options, board)],
-			this.getOptions(options),
-			board
-		))
+		const { board, threads } = parseThreadsResponse(response)
+		return threads.map(thread => this.createThreadObject(thread, options, { board }))
+	}
+
+	/**
+	 * Parses "get threads list" page API response.
+	 * @param  {any} response
+	 * @param  {object} [options]
+	 * @return {Thread[]}
+	 */
+	parseThreadsPage(response, options) {
+		const { board, threads } = parseThreadsPageResponse(response)
+		return threads.map(thread => this.createThreadObject(thread, options, { board }))
 	}
 
 	/**
@@ -58,32 +62,22 @@ export default class Makaba extends Engine {
 	 * @return {Thread}
 	 */
 	parseThread(response, options) {
-		const {
-			thread,
-			comments,
-			board
-		} = parseThreadResponse(response)
+		const { thread, board } = parseThreadResponse(response)
 		// Fix incorrect attachments count.
 		// https://gitlab.com/catamphetamine/imageboard/blob/master/docs/engines/makaba-issues.md
-		thread.attachmentsCount = comments.reduce((sum, comment) => sum += comment.files.length, 0)
-		return Thread(
-			thread,
-			comments.map(comment => this.parseComment(comment, options, board)),
-			this.getOptions(options),
-			board
-		)
+		thread.attachmentsCount = thread.comments.reduce((sum, comment) => sum += comment.files.length, 0)
+		return this.createThreadObject(thread, options, { board })
 	}
 
 	/**
-	 * Creates a `Comment` from comment data.
+	 * Parses comment data.
 	 * @param  {object} comment
 	 * @param  {object} options
-	 * @param  {object} board
-	 * @return {Comment}
+	 * @param  {object} parameters.board
+	 * @return {object}
 	 */
-	parseComment(comment, options, board) {
-		options = this.getOptions(options)
-		return Comment(parseComment(comment, options, board), options)
+	_parseComment(comment, options, { board }) {
+		return parseComment(comment, options, { board })
 	}
 
 	/**
@@ -98,9 +92,17 @@ export default class Makaba extends Engine {
 	/**
 	 * Parses "post" API response.
 	 * @param  {object} response
-	 * @return {boolean} Returns `true` if the vote has been accepted.
+	 * @return {number} Returns new thread ID or new comment ID.
 	 */
 	parsePostResponse(response) {
 		return parsePostResponse(response)
+	}
+
+	/**
+	 * Parses "report" API response.
+	 * @param  {object} response
+	 */
+	parseReportResponse(response) {
+		return parseReportResponse(response)
 	}
 }